requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
blinker==1.8.2
bokeh==3.4.1
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cssselect2==0.7.0
distro==1.9.0
Django==5.0.6
django-allauth==0.63.3
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-js-asset==2.2.0
django-paypal==2.1
djangorestframework==3.15.2
et-xmlfile==1.1.0
Flask==3.0.3
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
fonttools==4.53.1
fpdf==1.7.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.19.0
multidict==6.0.5
mysqlclient==2.2.4
numpy==2.0.1
openpyxl==3.1.4
oscrypto==1.3.0
packaging==24.1
pandas==2.2.2
paypal==1.2.5
pillow==10.3.0
plotly==5.22.0
pyaes==1.6.1
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydyf==0.11.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyMySQL==1.1.1
pypdf==4.2.0
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.21.1
python-http-client==3.3.7
python-telegram-bot==21.4
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
rsa==4.9
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.0
svglib==1.5.1
Telethon==1.36.0
tenacity==8.4.1
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.3
xhtml2pdf==0.2.16
XlsxWriter==3.2.0
xyzservices==2024.6.0
yarl==1.9.4
zopfli==0.2.3
